1-30-15
I create a txt file that is trying to explore the needed functions as the game is running. Take a look at it and add to it, but as you do try to think of what the game
needs to be doing. What are the actions that the game is going through the process of completing.